+++ application/main-app/src/main/java/org/mainapp/domain/post/controller/request/UpdatePostsMetadataRequest.java
@@ -13,7 +13,7 @@ import lombok.Getter;
 @Getter
 @AllArgsConstructor
 @Schema(description = "게시물 기타 정보 수정 API 요청 본문")
-public class UpdatePostsRequest {
+public class UpdatePostsMetadataRequest {
 
 	@Schema(description = "게시물 수정 정보 리스트")
 	@NotNull(message = "게시물 수정 정보를 입력해주세요.")
